Restructure settings (#481)
