+++ lib/core/widgets/images/analog_logo.dart
