+++ lib/features/occupation/presentation/widgets/occupation_form.dart
@@ -1,8 +1,8 @@
 import 'package:coffeecard/base/strings.dart';
+import 'package:coffeecard/core/widgets/list_entry.dart';
 import 'package:coffeecard/features/occupation/domain/entities/occupation.dart';
-import 'package:coffeecard/widgets/components/list_entry.dart';
+import 'package:coffeecard/features/settings/presentation/widgets/settings_list_entry.dart';
 import 'package:coffeecard/widgets/components/section_title.dart';
-import 'package:coffeecard/widgets/components/settings_list_entry.dart';
 import 'package:coffeecard/widgets/components/tickets/rounded_button.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
