+++ lib/features/receipt/presentation/widgets/receipt_card.dart
@@ -1,10 +1,10 @@
 import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/colors.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
+import 'package:coffeecard/core/widgets/images/analog_logo.dart';
 import 'package:coffeecard/utils/responsive.dart';
 import 'package:coffeecard/utils/time_since.dart';
 import 'package:coffeecard/widgets/components/card.dart';
-import 'package:coffeecard/widgets/components/images/analog_logo.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:intl/intl.dart';
