+++ lib/features/settings/presentation/widgets/faq.dart
@@ -1,25 +1,6 @@
-import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
-import 'package:coffeecard/widgets/components/scaffold.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
-
-class FAQPage extends StatelessWidget {
-  static Route get route => MaterialPageRoute(builder: (_) => FAQPage());
-
-  @override
-  Widget build(BuildContext context) {
-    return AppScaffold.withTitle(
-      title: Strings.faq,
-      body: ListView(
-        padding: const EdgeInsets.all(16),
-        children: Strings.faqEntries.entries
-            .map((e) => FAQ(question: e.key, answer: e.value))
-            .toList(),
-      ),
-    );
-  }
-}
 
 class FAQ extends StatefulWidget {
   const FAQ({required this.question, required this.answer});
