+++ lib/features/settings/presentation/widgets/settings_list_entry.dart
@@ -1,6 +1,6 @@
 import 'package:coffeecard/base/style/colors.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
-import 'package:coffeecard/widgets/components/list_entry.dart';
+import 'package:coffeecard/core/widgets/list_entry.dart';
 import 'package:flutter/material.dart';
 
 class SettingListEntry extends StatelessWidget {
