+++ lib/widgets/pages/splash/splash_loading_page.dart
@@ -1,5 +1,5 @@
 import 'package:coffeecard/base/style/colors.dart';
-import 'package:coffeecard/widgets/components/images/analog_logo.dart';
+import 'package:coffeecard/core/widgets/images/analog_logo.dart';
 import 'package:flutter/material.dart';
 
 class SplashLoadingPage extends StatelessWidget {
